Move CoreCLR-only serialization tests to a separate file
