+++ src/System.Runtime.Serialization.Xml/tests/System.Runtime.Serialization.Xml.Tests.csproj
@@ -18,7 +18,9 @@
     <Compile Include="Performance\XmlUTF8TextReaderTests.cs" />
     <Compile Include="Utils.cs" />
     <Compile Include="SerializationTypes.cs" />
+    <Compile Include="SerializationTypes.CoreCLR.cs" />
     <Compile Include="DataContractSerializer.cs" />
+    <Compile Include="DataContractSerializer.CoreCLR.cs" />
     <Compile Include="DataContractSerializerStressTests.cs" />
     <Compile Include="DataContractSerializerTestData.cs" />
     <Compile Include="MyResolver.cs" />
